+++ pb/sf/substreams/rpc/v2/substreams_test.go
@@ -1,4 +1,4 @@
-package pbsubstreamsrpc
+package pbsubstreamsrpcv2
 
 import (
 	"fmt"